// src/store/home/home.type.ts
export enum HomeActionsType {
  ADD_CATEGORIES = 'ADD_CATEGORIES',

  ADD_PRODUCT = 'ADD_PRODUCT',

  ADD_CART = 'ADD_CART',
  ADD_CART_SUCCESS = 'ADD_CART_SUCCESS',

  UPDATE_CART = 'UPDATE_CART',
  UPDATE_CART_SUCCESS = 'UPDATE_CART_SUCCESS',

  CLEAR_CART = 'CLEAR_CART',
  CLEAR_CART_SUCCESS = 'CLEAR_CART_SUCCESS',

  REMOVE_PRODUCT_CART = 'REMOVE_PRODUCT_CART',
  REMOVE_PRODUCT_CART_SUCCESS = 'REMOVE_PRODUCT_CART_SUCCESS',

  ERROR = 'ADD_CART_ERROR',
}